chore(Table): move utils to a more suitable location
