+++ packages/core/src/Table/TableRow/TableRow.styles.tsx
@@ -1,7 +1,22 @@
 import { createClasses } from "@hitachivantara/uikit-react-utils";
 import { theme } from "@hitachivantara/uikit-styles";
 
-import { getBorderStyles } from "../utils/utils";
+const getBorderStyles = (color: string, rowBorderRadius: string) => {
+  return {
+    "& td": {
+      borderTop: `1px solid ${color}`,
+      borderBottom: `1px solid ${color}`,
+    },
+    "& td:first-of-type": {
+      borderLeft: `1px solid ${color}`,
+      borderRadius: `${rowBorderRadius} 0 0 ${rowBorderRadius}`,
+    },
+    "& td:last-of-type": {
+      borderRight: `1px solid ${color}`,
+      borderRadius: `0 ${rowBorderRadius} ${rowBorderRadius} 0`,
+    },
+  };
+};
 
 export const { staticClasses, useClasses } = createClasses("HvTableRow", {
   /** Styles applied to the component root class. */
