+++ packages/core/src/Table/index.ts
@@ -7,4 +7,3 @@ export * from "./TableRow";
 export * from "./TableCell";
 export * from "./hooks";
 export * from "./renderers";
-export * from "./utils";
