+++ packages/core/src/Table/renderers/index.ts
@@ -1,13 +1,4 @@
-export {
-  hvTextColumn,
-  hvNumberColumn,
-  hvDateColumn,
-  hvExpandColumn,
-  hvTagColumn,
-  hvSwitchColumn,
-  hvDropdownColumn,
-  hvProgressColumn,
-} from "./renderers";
+export * from "./renderers";
 export * from "./DateColumnCell";
 export * from "./DropdownColumnCell";
 export * from "./ProgressColumnCell";
