--- packages/core/src/Table/utils/index.ts
@@ -1,6 +1,0 @@
-export {
-  hvStringFallback,
-  hvNumberFallback,
-  hvNodeFallback,
-} from "./fallbacks";
-export * from "./utils";
